Add isoform serialization; tested but unverified.
(Looks ok though.)

Changed scoring algorithm to:
        100*(matches)/(matches+mistakes)
where mistakes are insertions, deletions, and mismatches. This score is
a percent. If a feature is only partially covered, all uncovered bases
are counted as deletions; this prevents scenarios like a single correct
base overlapping an otherwise-empty feature, leading to a high
match-mistake ratio.